Rotating device will not reset game
Instead of passing the game id to GameActivity through an intent, the
game is saved to a file. If the continue saved game button is pressed
it goes to game activity without saving over the existing game. This
should allow greater flexibility in the game that is passed and should
allow for custom mode creation.