Fixed duplicate tiles bug
Going from GameActivity to another activity and back used to cause
tiles to duplicate.

If GameActivity is loaded for the first time the grid is created. If
the user returns to this activity after switching to another activity,
the grid is still recreated because there is a chance that android
killed this activity in the background