Workaround for game breaking bug
The game is not assigning a game mode id to a game when the app is
first launched. It attempts to get the game title which relies on the
mode id. The id is invalid and the app crashes. The game now shows 2048
as the game mode title if the id is invalid.