Added how to play animation
InfoActivity now says “Swipe to move the board. Tiles with the same
value combine. Reach a 2048 tile to win.” Below it is a VideoView of a
game making three moves demonstrating how to play. The video only takes
up 93 KB so I decided to having it play instead of moving the tiles in
real time.