Moved help and inventory to MainActivity
I moved the help button out of the overflow menu in SelectModeActivity
and into a button in MainActivity.
See #83 Move items out of menu overflow

The power up and undo inventory is now shown on MainActivity in
addition to SelectModeActivity.

The app logo on MainActivity is now using a larger image instead of
ic_launcher