Added dialog when reaching 2048 tile
When a 2048 tile is created a dialog is shown saying congratulations
with an image view of a 2048 tile. There are two buttons continue
playing which dismisses the dialog and brag which creates a share
intent. Tapping on the share button unlocks the achievement “Brag to
your friends”.

See #89 Add back share action provider