Multiplayer game does not overwrite current game
Starting a multiplayer game would require creating a new game, setting
multiplayer active in that game to true, saving the game to the file
system, overwriting any current save, and loading it from game fragment.

Now when the game fragment is created it checks if its parent is
MultiplayerActivity. It places that value into a boolean which game
fragment will check.