Tiles on GameFragment scale to fit the screen
Only if the width >= height. If the height is larger than the width
tiles go off the bottom of the screen. Tiles need to be scaled every
time one is created so I might make a map of the scaled images to speed
up the game.

Also increased the max width and height for custom games to 20x20 and
fixed a bug where the height and width of the preview was switched.

See #63 Scale grid and tiles to fit screen
See #120 Show preview of custom game