Moved sign in/out buttons to MainActivity
The sign in/out buttons are now above play games layout instead of in
it. The sign out button is now represented as the G+ logo. Tapping on
it shows a dialog saying you are logged in and provides a button to log
out. The sign in button now actually works. The game now only signs in
automatically 2 times instead of 3.

Closes #122 Add sign out of Google Play Games button