Landscape layouts for main/game activities
In main activity the ad is thin across the entire bottom. Play games
layout is on the right side of the screen vertically.

In game activity the restart button is in the top right. The game is to
the left of it. All other views are to the left of the game.

The dialog to create a new game is not formatted for horizontally yet
and the bottom is cut off.

See #128 Support landscape mode