Code cleanup
Also changed the continue game yes/no buttons to confirmation buttons
built into the dialog. Also added a background to the sign out button.

Moved many more strings to xml. This is all of the hardcoded strings
that the inspector found but there are probably a few more.

See #58 Move strings to xml