Minor layout changes
Removed the to game button from MultiplayerActivity.
Game modes in the new game dialog have a little bit of padding to
prevent going on top of the border.
Changed the icon for the sign out button to a rectangular black tile
instead of a square.
You need to be signed in to switch to MultiplayerActivity.