New color scheme for GameFragment
The top of the game containing the mode, score, high score, time left,
and restart button is light blue. The bottom of the screen with undo,
power up, active attacks, and the game is a darker blue. The blank
tiles behind the game is now light blue instead of orange.

Also increase the text size of the time left. 50sp → 60sp

Increased the version number 1.0.1 → 1.0.2 and generated a new signed
apk.