Multiplayer attacks become pending if one is active
Previously when attack is received when another one is active it
becomes lost. Now if the attack is the same one as currently active the
attack duration is increased. If it is a different attack it gets added
to an array list of pending attacks.

Also increased the multiplayer version 39 → 40. This makes it
incompatible with previous versions.