New layout for custom game activity
The top is containing the grid size and game modes is now light blue.
The bottom with the game preview and start game button is blue. The
“Create Game” button has been changed to “Start” and the background of
it has also been changed from yellow and orange to blue. Also, the game
modes checkboxes are centered under Game Modes instead of aligned to
the right of the screen.