Impossible to not be able to move in multiplayer
If you move and the game is lost the board shuffles. If you still
cannot move the lowest value tile is removed from the board. It may
still be possible to not be able to move, I need to test this more.

See #160 Make it impossible to not be able to move in multiplayer